% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replxx.R
\name{replxx}
\alias{replxx}
\alias{replxx_set_highlighter}
\alias{replxx_set_completion}
\alias{replxx_add_completions}
\alias{replxx_set_hinter}
\alias{replxx_add_hints}
\alias{replxx_input}
\alias{replxx_print}
\alias{replxx_set_preload}
\alias{replxx_history_add}
\alias{replxx_history_size}
\alias{replxx_set_word_break_characters}
\alias{replxx_set_special_prefixes}
\alias{replxx_set_max_line_size}
\alias{replxx_set_max_hint_rows}
\alias{replxx_set_double_tab_completion}
\alias{replxx_set_complete_on_empty}
\alias{replxx_set_beep_on_ambiguous_completion}
\alias{replxx_set_no_color}
\alias{replxx_set_max_history_size}
\alias{replxx_get_history_line}
\alias{replxx_save_history}
\alias{replxx_load_history}
\alias{replxx_clear_screen}
\title{Low level interface for replxx}
\usage{
replxx()

replxx_set_highlighter(replxx, fun)

replxx_set_completion(replxx, fun)

replxx_add_completions(completions, strings)

replxx_set_hinter(replxx, fun)

replxx_add_hints(hints, strings)

replxx_input(replxx, prompt = "? ")

replxx_print(replxx, fmt, ...)

replxx_set_preload(replxx, text)

replxx_history_add(replxx, lines)

replxx_history_size(replxx)

replxx_set_word_break_characters(replxx, chars)

replxx_set_special_prefixes(replxx, specials)

replxx_set_max_line_size(replxx, len)

replxx_set_max_hint_rows(replxx, count)

replxx_set_double_tab_completion(replxx, on)

replxx_set_complete_on_empty(replxx, on)

replxx_set_beep_on_ambiguous_completion(replxx, on)

replxx_set_no_color(replxx, on)

replxx_set_max_history_size(replxx, len)

replxx_get_history_line(replxx, index)

replxx_save_history(replxx, path)

replxx_load_history(replxx, path)

replxx_clear_screen(replxx)
}
\description{
Low level interface for replxx
}
